TranslateDemo/requirements.txt
azure.ai.translation.text
azure.identity
